Saner sorting behavior in documents
Old BrickStore versions did not have a model/view architecture, so sorting
directly influenced the order in which items were internally held and also
in which they were saved.
The new version had a weird mix of a modern model/view architecture and old
code excpecting the model to be sorted.

This patch will change this behavior:
- Every document has a natural sort order (the order in which items are
  added to the document) and this will never change.
- When you load a document, the table view is not sorted (no sort
  indicator) which gives you that natural order.
- There is a new (leftmost) column called "Index" which you can use to go
  back to the natural order at any time after sorting differently.
- Adding items to a sorted document will add the new items at the bottom
  (just like in the old 1.x versions). Click the header to re-sort the
  list afterwards.
- If you don't care about his "Index" column, then just hide it.

Also modernized the code a bit using std::initializer_list.