Added back support for saving column layouts
By default there's the built-in "Brickstore default" and a configurable "User default" which is used for newly opened documents.
The user can save any number of custom, named layouts on top of that.

The UI for renaming and deleting these custom layouts is not finished yet.

Layouts are not saved per document yet.