Per document classic filters
This should make it easier for non-keyboard warriors to work with filters.
Also consolidated all the column layout in a single setup function to
avoid missing bits and pieces when adding new columns (this change also
required the move to C++20 in order to use designated member initializers).

Downsides: filters cannot be saved at the moment and they are not accesible
from the scripting API.

Closes: #408
Closes: #379
Closes: #268
Closes: #258
Closes: #52
Closes: #1