Prepare the move from the git master to main branch
We have to do this because I broke the update check in 2021.10.2: these
versions will just crash, if CHANGELOG.md in the master branch shows
that a newer version is available.
The only way to get somewhat cleanly out of this mess is to abandon
master, move development to main, notify people repeatedly via
the Announcments mechanism to update and finally retire the master
branch in a few months.