Prevent crash on db update with active picture/priceguide downloads
Cancelling all outstanding transfers cannot be done synchronously
anymore due to the number of threads involved.

Ideally there needs to be an async DB lock protocol, but for the
time being we're counting on the actual DB download to take long
enough for Qt's event queue to clear.
There is a very slim chance that this might not be enough, so we
make sure to not delete, but leak, the still locked resources.