Fix potential crash after updating the DB
In this specific instance, the user had used the Add Items dialog to
add parts to a document, then closed all open documents and then
updated the database. This would not re-create the Add Item dialog.
Whenever the update timer for the history update came in, BrickStore
would crash because the history entries still had pointers to the old
DB.

Sentry-Issue: BRICKSTORE-87