Keep better track of ref'ed Q_GADGETs on the QML side
If priceGuideUpdated() was connected to a QML signal handler, we would
"loose" a QmlPriceGuide object per signal emission to the JS heap.
Although these gadgets would be garbage collected at some point, they
are ref'ing the corresponding C++ side objects and preventing a cache
clear when updating the database (and eventually leading to a crash).

Also modernized the code a bit.