Detect item-id changes during DB updates and update inventories
If an item-id changes, the inventories containing this item are not
marked as updated. We now use the changelog to see if an item-id
has changed since the last time we downloaded the inventory.

This is not 100% perfect, because the timestamps are only in days,
but it's way better than not checking this at all.