Fix crash when duplicating more than one lot
This is a nasty side-effect of not delaying the selection update
anymore: we were updating the const ref to the list of selected lots,
while iterating over it.

The simple fix would be to just copy the list of selected lots, but
this leads to a severe performance drop when duplicating a huge
number of lots, because the selectionUpdated signal is emitted for each
and every duplicated lot.
In addition to a new insertLotsAfter overload for duplicate, the
updateSelection function was also heavily optimized to not rely on
high-level QItemSelection code, but to assume full-row selections
in our use case.

Sentry-issue: BRICKSTORE-FM