Redo the GPU blacklisting
Just not rendering seems to be enough for Intel and AMD GPUs, but old
NVIDIAs already crash when initializing an empty Quick3D view.

In order to have the GPU name available early during startup, the
SystemInfo initialization had to be moved back into the main thread.
This will slow down the startup on macOS and Linux slightly.

Closes: #621