Workaround for the macOS QSystemLocale's toString() being very slow
This should be fixed in Qt by cacheing the result of
QLocaleData::query(), but until this lands upstream, we can work
around this by setting the default locale to a Qt one, based on
the current system locale.
This will not perfectly mirror things like the exact date and
time format from the macOS system settings, but at least it makes the
filter usable again on large documents.

Closes: #641