Fix broken state when closing a document in a split-off document
If you have one document, then create a split into a new window and
then close that document in the split-off window, the mainwindow would
correctly switch to the welcome screen, but immediately switch back
to a broken, empty document view.
This is caused by the window activation of the mainwindow happening in
parallel, leading to a widget focus change, which confused BrickStore.