Correctly implement ChangeLog handling for item and color ids
The old changelog handling did not take into account that the changelog
had to be applied from oldest to newest to correctly catch id swaps.
This was also due to the fact, that we didn't know where to start in
the changelog history. Adding the recursive resolve to this broken
setup ended in infinte loops on some item renames (e.g. 72206)

Newly saved BSX files now carry the latest BL changelog id to correctly
resolve any future renames. For old BSX files, as well as BrickLink XML
files, the file modification time will be used as a starting point in
the changelog. This is just a best guess though, because BrickLink's
temporal precision for changelog entries is a day, plus the BrickStore
database is typically a few days older than the file being saved.
It does work out quite well in practice though.

Closes: #686