Add support for item relationships
Downloading the relations is a bit of hack, because BL does not offer
an XML download API. Instead we are scraping the HTML pages with
regular expressions, which is bound to break sooner or later.

Showing the relations in the UI is done via the inventory widget
(appears-in/consists-of/...). Right now the actual type of relation
is not yet shown, as this would need a redesign of the current model
to support hierarchies.

Closes: #67