Save orders in a SQLITE database
Instead of saving to the filesystem as XML files, we now store all
orders in a SQLITE database.
This solves the immediate problem of the order date not being constant
when batches are involved. Long term, the import order dialog can be
optimized to not load all orders from the get-go, but instead use a
QSqlQueryModel.

Closes: #687