Fix color-lock interfering with go-to-item and the browse history
Using "go to item" from a color-locked item to another item tries to
keep the color lock, if the new item is available in that color.
Otherwise the color-lock will be disabled.

In addition the last selected color and lock status is now replayed
correctly when using the browse history.

Closes: #696