Replace all vectors in the database with with PooledArrays
We are saving (16 - sizeof(T)) bytes per array, which amounts to
roughly 20MB at runtime.
In addition we are now using a monotonic pool allocator, so all of
the small string and index table allocations (37MB in total) are
allocated/freed in 1MB steps, tremendously speeding up the database
loading and updating.