Fix massive slow-down in the appears-in view on popular items
Computing the appars-in view for items that appear in a lot of sets
(e.g. Plate 1 x 2) would take in the order of seconds.
The problem here is the non-uniform row heights take came in through
the section headers in the relationships view, and these take a lot
of time to compute.
We now use non-uniform row heights for the relationships model only:
it's still slow, but (a) these models are fairly small and (b) we
DO want the section header size independent from the item size.