Make the MenuCombobox behave on macOS
The popup menu placement on macOS was very weird in on non-primary
screens. Instead of duplicating Qt's placement logic, we just use
Qt's internal popup on macOS.
As palettes are inherited by child widgets, we have to split off the
gradient on the document selector into its own background widget.

Closes: #708