Improve camera dialog open times, which got really bad on Windows
The actual time to enumerate the cameras seems to depend on the OS and
driver versions, so we have to deal with it.

The solution is to keep an instance of the ItemScannerDialog alive
at all times, but we have to now deal with updates to the OSes list of
camera devices.
On top of that, the video overlay window has a bug, were the video
is lost after a hide/show cycle. We can get around that with a dummy
resize though to force a refresh.

Closes: #792
Closes: #760