Revert adding an organization name to the QCoreApplication
This was done for 2023.11.2 to decouple the cookie saving from the
Config mechanism, but this ended up backfiring:

Both the 'Cache' and 'AppData' standard paths were affected by that
change (as in, there's an additional sub-directory named "BrickStore").

The Cache part is easy, as we can just go back to the old
cache directory (it's just a cache after all). For the AppData, we
need to make sure that people haven't copied and later modified
their extensions.