Complete database generator overhaul
Instead of keep a gazillion files spread all over the cache directory,
we use single ZIP file to store all the input for a generation run
(which makes this 100% reproducible) and we also use the file as a
cache for the inventories on subsequent runs.

Also skip-download now works perfectly fine in full offline mode,
because we store every single downloaded file in that ZIP archive.

This should make it possible to finally run the generator as a
GitHub CI workflow.