Take both app state and window visibility in account for camera state
The new "(de)activate camera if app is (in)active" was too simplistic.
We now track both the application foreground state and the preview
window visibility to control the camera activation.

This patch also gets rid of the "deactivate camera 10sec after the
preview window lost focus" completely, as this complicates the
state machine a lot for very little gain.

Closes: #878