Rewrite the threaded RenderWidget part loading
There are a lot of Sentry reports of crashes in this area, and they
all point towards the old updateGeometries() co-routine being called
again while it was awaiting the geometry calculations in another
thread.

The new version uses a more robust setup of 2 then-ables (loading and
geometry calculation), but both sync back to the main thread before
continuing. This also makes it easy to skip the geometry calculation,
if the user already selected a different part in the meantime.